04_后端开发/03_业务代码/finance/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import OrderViewSet, RefundRecordViewSet

router = DefaultRouter()
router.register(r'orders', OrderViewSet, basename='order')
router.register(r'refunds', RefundRecordViewSet, basename='refund')

urlpatterns = [
    path('', include(router.urls)),
]