04_后端开发/03_业务代码/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
redis>=4.5
celery>=5.3
PyJWT>=2.8
python-dotenv>=1.0
Pillow>=10.0
openpyxl>=3.1
django-filter>=3.0
markdown>=3.5
django-redis>=5.0
gunicorn>=21.0